+++ __test__/date.spec.js
@@ -205,4 +205,30 @@ describe('date.js', () => {
             pure.definitions.date.weekday.abbr_context = backupAbbrContext;
         });
     });
+
+    describe('birthDay()', () => {
+        it('return generated birthday', () => {
+            const date = pure.date.birthDay();
+
+            assert.ok(date < new Date())
+        });
+
+        it('return generated birthday given min and max age', () => {
+            const date = pure.date.birthDay(10, 12);
+            const actual = (new Date()).getFullYear()
+            const generated = (new Date(date)).getFullYear()
+
+            assert.ok(10 <= (actual - generated))
+            assert.ok(12 >= (actual - generated))
+        });
+
+        it('return generated birthday given min and max age inverted', () => {
+            const date = pure.date.birthDay(12, 10);
+            const actual = (new Date()).getFullYear()
+            const generated = (new Date(date)).getFullYear()
+
+            assert.ok(10 <= (actual - generated))
+            assert.ok(12 >= (actual - generated))
+        });
+    });
 });